feat: merge formats and state
